Refactoring out personalization functionality
* and a little cleanup
